add feature flags: `sqlite` and `postgres`
Enables sqlxinsert to be used without a dependencie on sqlite and postgres.
Which is especially useful for users only interested in the sqlite part or the postgres part.
Both of the features are enabled by default and therefore are not breaking any working code.

This was tested only for sqlite. Because of me not having a postgres instance avaiable.
```sh
cargo test --no-default-features --features sqlite
```

The new features are names `sqlite` and `postgres` following the naming scheme of sqlx.